Fix font detection for newer macOS.
Newer versions of macOS are now getting Python from the Xcode command
line tools package and no longer contains the Cocoa package.  Turns out
there's another tool, `atsutil`, that can be used to get a font list.
This tool has been around since OS X 10.5.  Since we're well beyond that
support window, let's drop the Cocoa based technique and use `atsutil`
to get the listing.  It's not the friendliest output, so it required a
little massaging to make the listing usable, but it's fast works well.